requirements.txt
alembic==1.15.1
appdirs==1.4.4
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email-validator==2.1.0.post1
Flask==2.0.1
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
# Flask-Bcrypt
Flask-Bcrypt==1.0.1

greenlet==3.1.1
idna==3.10
imgkit==1.2.3
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
python-dotenv==1.0.1
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
SQLAlchemy==1.4.46
stripe==7.11.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
websockets==10.4
Werkzeug==2.0.1
WTForms==3.2.1
zipp==3.21.0
numpy>=2.0.0,<3.0.0
boto3==1.37.27
dropbox==12.0.2
pandas>=1.5.3
annotated-types==0.7.0
blis>=1.3.0
catalogue==2.0.10
cloudpathlib==0.21.0
confection==0.1.5
cymem==2.0.11
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
murmurhash==1.0.12
Pillow==10.0.0
preshed==3.0.9
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
rich==14.0.0
scikit-learn>=1.6.1
scipy>=1.11.4
setuptools==80.0.0
shellingham==1.5.4
smart-open==7.1.0
spacy>=3.8.5
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
srsly>=2.5.1
thinc>=8.3.6
threadpoolctl==3.6.0
typer==0.15.3
typing-inspection==0.4.0
tzdata==2025.2
wasabi==1.1.3
weasel==0.4.1
wheel==0.45.1
wrapt==1.17.2
PyMuPDF==1.25.5
APScheduler==3.11.0
xlsxwriter==3.2.3
openpyxl==3.1.5
fpdf2==2.8.3
plotly==6.2.0
pyodbc==5.2.0
cryptography==45.0.6
matplotlib==3.10.5
seaborn==0.13.2